# api-spec.yml
openapi: 3.0.3
info:
  title: Book Management API
  description: Book Management API
  termsOfService: http://swagger.io/terms/
  version: "1"
servers:
  - url: http://localhost:3000
paths:
  /books:
    post:
      tags:
        - Book
      summary: Add a book
      description: Add a book
      parameters:
        - name: api-version
          in: header
          description: Version of this API. Different version numbers represent API changes that are incompatible with each other.
          required: true
          schema:
            maximum: 1
            minimum: 1
            type: integer
            default: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
      responses:
        201:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        400:
          description: __Bad request.__
          headers:
            content-length:
              schema:
                type: integer
                format: long
            content-type:
              schema:
                maxLength: 64
                minLength: 1
                type: string
                x-valid-chars: ^[a-zA-Z0-9=/-;]+$
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationOutcome'
        422:
          description: __Unprocessable Entity.__
          headers:
            content-length:
              schema:
                type: integer
                format: long
            content-type:
              schema:
                maxLength: 64
                minLength: 1
                type: string
                x-valid-chars: ^[a-zA-Z0-9=/-;]+$
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationOutcome'
        500:
          description: __Internal Server Error.__
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationOutcome'
    get:
      tags:
        - Book
      summary: Returns all books with matching criteria
      description: Returns all books with matching criteria
      parameters:
        - name: api-version
          in: header
          description: Version of this API. Different version numbers represent API changes that are incompatible with each other.
          required: true
          schema:
            maximum: 1
            minimum: 1
            type: integer
            default: 1
        - name: genre
          in: query
          description: Genre of books to be retrived.
          schema:
            maxLength: 36
            minLength: 1
            type: string
          x-valid-chars: '[a-zA-Z0-9-_.]'
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetBook'
        400:
          description: __Bad request.__
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationOutcome'
        500:
          description: __Internal Server Error.__
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationOutcome'
  /books/{bookId}:
    get:
      tags:
        - Book
      summary: Returns book by the provided book ID
      description: Retrives book by provided ID
      parameters:
        - name: api-version
          in: header
          description: Version of this API. Different version numbers represent API changes that are incompatible with each other.
          required: true
          schema:
            maximum: 1
            minimum: 1
            type: integer
            default: 1
        - name: bookId
          in: path
          description: ID with which book should be retrieved.
          required: true
          schema:
            maxLength: 36
            minLength: 1
            type: string
          x-valid-chars: '^[0-9a-fA-F]{24}$'
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        400:
          description: __Bad request.__
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationOutcome'
        404:
          description: Book not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        500:
          description: __Internal Server Error.__
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationOutcome'
    put:
      tags:
        - Book
      summary: Update genre of book
      description: Update genre of book
      parameters:
        - name: api-version
          in: header
          description: Version of this API. Different version numbers represent API changes that are incompatible with each other.
          required: true
          schema:
            maximum: 1
            minimum: 1
            type: integer
            default: 1
        - name: bookId
          in: path
          description: id with which book should be retrieved.
          required: true
          schema:
            maxLength: 36
            minLength: 1
            type: string
          x-valid-chars: '^[0-9a-fA-F]{24}$'
      requestBody:
        description: Update genre for existing book in store
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
      responses:
        201:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        400:
          description: __Bad request.__
          headers:
            content-length:
              schema:
                type: integer
                format: long
            content-type:
              schema:
                maxLength: 64
                minLength: 1
                type: string
                x-valid-chars: ^[a-zA-Z0-9=/-;]+$
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationOutcome'
        404:
          description: Book not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        500:
          description: __Internal Server Error.__
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationOutcome'
    delete:
      tags:
        - Book
      summary: Delete book by ID
      description:  Delete book by ID
      parameters:
        - name: api-version
          in: header
          description: Version of this API. Different version numbers represent API changes that are incompatible with each other.
          required: true
          schema:
            maximum: 1
            minimum: 1
            type: integer
            default: 1
        - name: bookId
          in: path
          description: id with which book should be retrieved.
          required: true
          schema:
            maxLength: 36
            minLength: 1
            type: string
          x-valid-chars: '^[0-9a-fA-F]{24}$'
      responses:
        204:
          description: __No Content.__
        400:
          description: __Bad request.__
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationOutcome'
        404:
          description: ___Book already deleted.__
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationOutcome'
        500:
          description: __InternaL Server Error.__
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationOutcome'
components:
  schemas:
    Book:
      title: Book
      required:
        - resourceType
        - title
        - author
        - isbn
      type: object
      properties:
        resourceType:
          type: string
          default: Book
          enum:
            - Book
        id:
          maxLength: 64
          minLength: 1
          type: string
          description: (server assigned) Logical ID of the resource.
          readOnly: true
          x-valid-chars: ^[a-zA-Z0-9-]+$
        title:
          maxLength: 1024
          minLength: 1
          type: string
          description: Title of the book
          default: "Book Title"
          x-valid-chars: ^[a-zA-Z0-9-_. ]+$
        author:
          maxLength: 1024
          minLength: 1
          type: string
          description: Name of the book
          default: "Book Author"
          x-valid-chars: ^[a-zA-Z0-9-_. ]+$
        genre:
          maxLength: 1024
          minLength: 1
          type: string
          description: Name of the book
          default: Common
          enum:
            - Common
            - Thriller
            - Horror
          x-valid-chars: ^[a-zA-Z0-9-_. ]+$
        isbn:
          maxLength: 1024
          minLength: 1
          type: string
          description: Name of the book
          default: "Book unique ISBN"
          x-valid-chars: ^[a-zA-Z0-9-_. ]+$
    GetBook:
      type: array
      minItems: 1
      maxItems: 5
      items:
        $ref:  '#/components/schemas/Book'
    OperationOutcome:
      properties:
        resourceType:
          type: string
          default: OperationOutcome
          enum:
            - OperationOutcome
          description: OperationOutcome in a failure case.
        errorType:
          type: string
        message:
          type: string
      required:
        - resourceType
        - errorType
        - message